Camera and some warnings fix

+++ Infinite City - Source Code/code/src/main.cpp
@@ -52,7 +52,7 @@ int main(int argc, char* argv[])
         std::cerr << "Failed to create GLFW window" << std::endl;
         glfwTerminate();
         return -1;
-    }
+    } 
     glfwMakeContextCurrent(window);
 
     // Initialize GLEW
@@ -69,7 +69,7 @@ int main(int argc, char* argv[])
 
     
     Camera mainCamera(camSpeed); // The main camera that will 
-    long infiniteCitySeed = time(NULL); // The overall starting seed of the city - This seed will entirely define the uniqueness of the looks of the city.
+    long infiniteCitySeed = static_cast<long>(time(NULL)); // The overall starting seed of the city - This seed will entirely define the uniqueness of the looks of the city.
     int cityStartingLength = 5; // How many blocks long is the city at the start of the program (before the city grows as we move around)?
     int cityStartingWidth = 5;  // How many blocks wide is the city at the start of the program (before the city grows as we move around)?
     int blockSize = 5; // Size of each block in the city, a block will contain elements on it such as buildings, parks, trees, etc. A building block of the city.
